feat: add producible products and production suggestions endpoints with DTOs and service methods

+++ backend/src/main/java/com/projedata/inventory_manager/dto/productMaterial/ProductMaterialDTO.java
@@ -3,7 +3,6 @@ package com.projedata.inventory_manager.dto.productMaterial;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 
-import java.math.BigDecimal;
 
 public record ProductMaterialDTO(
 
@@ -12,6 +11,6 @@ public record ProductMaterialDTO(
 
         @NotNull(message = "Required quantity cannot be null")
         @Positive(message = "Required quantity must be positive")
-        BigDecimal requiredQuantity
+        Integer requiredQuantity
 ) {
 }
